requirements.txt
google-adk>=0.1.0
python-dotenv>=1.0.1
pyttsx3>=2.90
openai-whisper>=20231117
torch>=2.1.0
